+++ components/catalogs/front/jest.config.js
@@ -9,5 +9,4 @@ module.exports = {
             ...require('./jest.integration'),
         },
     ],
-    workerIdleMemoryLimit: '300MB'
 };