+++ src/Akeneo/Category/front/src/feature/hooks/useEditCategoryForm.ts
@@ -5,16 +5,11 @@ import {NotificationLevel, useNotify, useRouter, useTranslate} from '@akeneo-pim
 import {saveEditCategoryForm} from '../infrastructure';
 import {useCategory} from './useCategory';
 import {EditCategoryContext} from '../components';
-import {
-  buildCompositeKey,
-  Attribute,
-  CategoryAttributeValueData,
-  CategoryPermissions,
-  EnrichCategory,
-  Template,
-} from '../models';
+import {buildCompositeKey, Attribute, CategoryAttributeValueData, EnrichCategory, Template} from '../models';
 import {alterPermissionsConsistently, categoriesAreEqual, populateCategory} from '../helpers';
 import {useTemplateByTemplateUuid} from './useTemplateByTemplateUuid';
+import {CategoryPermissions} from '../models/CategoryPermission';
+import {UserGroup} from './useFetchUserGroups';
 
 const useEditCategoryForm = (categoryId: number) => {
   const router = useRouter();
@@ -110,19 +105,25 @@ const useEditCategoryForm = (categoryId: number) => {
         return;
       }
 
+      if (Array.isArray(categoryEdited.properties.labels) && !categoryEdited.properties.labels.length) {
+        categoryEdited['properties']['labels'] = {
+          [localeCode]: label,
+        };
+      }
+
       setCategoryEdited(set(['properties', 'labels', localeCode], label, categoryEdited));
     },
     [categoryEdited]
   );
 
-  const onChangePermissions = (type: keyof CategoryPermissions, values: number[]) => {
+  const onChangePermissions = (userGroups: UserGroup[], type: keyof CategoryPermissions, values: number[]) => {
     if (categoryEdited === null) {
       return;
     }
 
-    const consistentPermissions = alterPermissionsConsistently(categoryEdited.permissions, {type, values});
+    const consistentPermissions = alterPermissionsConsistently(userGroups, categoryEdited.permissions, {type, values});
 
-    setCategoryEdited(set(['permissions'], consistentPermissions, categoryEdited));
+    setCategoryEdited(set('permissions', consistentPermissions, categoryEdited));
   };
 
   const onChangeAttribute = useCallback(