+++ src/Akeneo/Category/front/src/feature/models/Category.ts
@@ -1,7 +1,8 @@
-import {LabelCollection, ChannelCode, LocaleCode} from '@akeneo-pim-community/shared';
+import {ChannelCode, LabelCollection, LocaleCode} from '@akeneo-pim-community/shared';
 import {TreeNode} from './Tree';
 import {CompositeKeyWithoutLocale} from './CompositeKey';
 import {CategoryAttributeType} from './Attribute';
+import {CategoryPermissions} from './CategoryPermission';
 
 export type Category = {
   id: number;
@@ -23,12 +24,6 @@ export type EnrichCategory = {
 export type CategoryProperties = {
   code: string;
   labels: LabelCollection;
-};
-
-export type CategoryPermissions = {
-  view: number[];
-  edit: number[];
-  own: number[];
 };
 
 export interface CategoryAttributes {
@@ -81,6 +76,7 @@ export type CategoryTreeModel = {
   productsNumber?: number;
   templateLabel?: string;
   templateUuid?: string;
+  templateCode?: string;
 };
 
 export type FormField = {
@@ -145,6 +141,7 @@ const buildTreeNodeFromCategoryTree = (
   return {
     identifier: categoryTree.id,
     label: categoryTree.label,
+    code: categoryTree.code,
     childrenIds: Array.isArray(categoryTree.children) ? categoryTree.children.map(child => child.id) : [],
     data: categoryTree,
     parentId: parent,